Move translateAppStrings function to language_and_audio directory
