+++ apps/design/backend/src/language_and_audio/index.ts
@@ -1,3 +1,4 @@
+export * from './app_strings';
 export * from './election_strings';
 export * from './speech_synthesizer';
 export * from './translator';
